fix(sdk-python): eliminate sys.path hack for FlatBuffers imports
Move generated files from _wire/generated/pyrat/protocol/ to
_wire/protocol/ and post-process imports from pyrat.protocol →
pyrat_sdk._wire.protocol. Standard Python imports just work,
no sys.path manipulation needed.
