+++ sdk-python/pyproject.toml
@@ -44,15 +44,12 @@ lint.ignore = [
     "B008",      # Do not perform function call in argument defaults
 ]
 
-exclude = ["pyrat_sdk/_wire/generated"]
+exclude = ["pyrat_sdk/_wire/protocol"]
 
 [tool.ruff.lint.per-file-ignores]
 "tests/*" = ["PLC0415", "PLR2004", "PLR0915"]
-# Side-effect import (pyrat_sdk._wire) must precede pyrat.protocol imports for sys.path setup.
-# FlatBuffers boilerplate is inherently complex.
-"pyrat_sdk/_wire/codec.py" = ["I001", "C901"]
-"pyrat_sdk/bot.py" = ["I001", "C901", "PLR0912", "PLR0915"]
-"pyrat_sdk/options.py" = ["I001"]
+"pyrat_sdk/_wire/codec.py" = ["C901"]
+"pyrat_sdk/bot.py" = ["C901", "PLR0912", "PLR0915"]
 
 [tool.ruff.format]
 quote-style = "double"
