+++ sdk-python/pyrat_sdk/_wire/protocol/BotMessage.py
