+++ sdk-python/pyrat_sdk/_wire/protocol/BotPacket.py
