+++ sdk-python/pyrat_sdk/_wire/protocol/Direction.py
