+++ sdk-python/pyrat_sdk/_wire/protocol/Identify.py
@@ -45,7 +45,7 @@ class Identify(object):
             x = self._tab.Vector(o)
             x += flatbuffers.number_types.UOffsetTFlags.py_type(j) * 4
             x = self._tab.Indirect(x)
-            from pyrat.protocol.OptionDef import OptionDef
+            from pyrat_sdk._wire.protocol.OptionDef import OptionDef
             obj = OptionDef()
             obj.Init(self._tab.Bytes, x)
             return obj
