+++ sdk-python/pyrat_sdk/_wire/protocol/Player.py
