+++ sdk-python/pyrat_sdk/_wire/protocol/Stop.py
