+++ sdk-python/pyrat_sdk/options.py
@@ -24,7 +24,6 @@ from __future__ import annotations
 import sys
 from typing import Any
 
-
 # ---------------------------------------------------------------------------
 # Base descriptor
 # ---------------------------------------------------------------------------
